styles : responsive search
