+++ src/components/NavbarMobileSearch/NavbarMobileSearch.css
@@ -1,0 +1,12 @@
+.mobile-search{
+padding: 5px;
+border-radius: 0;
+}
+.mobile-search-bar{
+    background-color: var(--secondary-color);
+    border: 1px solid var(--secondary-color);
+    margin-right: auto;
+    margin-left: 2rem;
+}
+
+
